backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
beautifulsoup4
langchain-google-genai>=1.0.7
google-generativeai>=0.7.0
langchain>=0.2.0
langchain-core>=0.2.0
python-dotenv
requests